can generate invoices

+++ backend/src/main/java/com/hotel/payments/entity/BillEntity.java
@@ -4,13 +4,9 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.Inheritance;
-import javax.persistence.InheritanceType;
-import javax.persistence.JoinColumn;
 import javax.persistence.OneToOne;
 
 import com.hotel.payments.interfaces.IBilling;
-import com.hotel.payments.interfaces.IChargeable;
 import com.hotel.payments.interfaces.IRefundable;
 
 import lombok.Data;
@@ -21,7 +17,7 @@ public class BillEntity implements IRefundable, IBilling {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int billId;
+    private Long billId;
 
     @OneToOne
     private PaymentEntity payment;
